+++ src/Contract/SqlTwigInterface.php
@@ -4,20 +4,27 @@ declare(strict_types=1);
 
 namespace Zjk\SqlTwig\Contract;
 
+use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Cache\QueryCacheProfile;
 use Doctrine\DBAL\Driver\Exception as ExceptionDriver;
 use Doctrine\DBAL\Exception;
+use Doctrine\DBAL\ParameterType;
 use Doctrine\DBAL\Result;
 use Doctrine\DBAL\TransactionIsolationLevel;
+use Doctrine\DBAL\Types\Type;
 use Twig\Error\LoaderError;
 use Twig\Error\RuntimeError;
 use Twig\Error\SyntaxError;
 
+/**
+ * @psalm-type WrapperParameterType = string|Type|ParameterType|ArrayParameterType
+ * @psalm-type WrapperParameterTypeArray = array<int<0, max>, WrapperParameterType>|array<string, WrapperParameterType>
+ */
 interface SqlTwigInterface
 {
     /**
      * @param array<string, mixed> $args
-     * @param array<string, int>   $types
+     * @psalm-param WrapperParameterTypeArray $types
      *
      * @throws ExceptionDriver
      * @throws Exception
@@ -27,8 +34,5 @@ interface SqlTwigInterface
      */
     public function executeQuery(string $queryPath, array $args = [], array $types = [], ?QueryCacheProfile $qcp = null): Result;
 
-    /**
-     * @param TransactionIsolationLevel::* $transactionIsolationLevel
-     */
-    public function transaction(\Closure $func, int $transactionIsolationLevel = TransactionIsolationLevel::READ_COMMITTED): ?Result;
+    public function transaction(\Closure $func, TransactionIsolationLevel $transactionIsolationLevel = TransactionIsolationLevel::READ_COMMITTED): ?Result;
 }