+++ src/DependencyInjection/Extension.php
@@ -7,7 +7,7 @@ namespace Zjk\SqlTwig\DependencyInjection;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\XmlFileLoader;
-use Symfony\Component\HttpKernel\DependencyInjection\Extension as SymfonyExtension;
+use Symfony\Component\DependencyInjection\Extension\Extension as SymfonyExtension;
 
 class Extension extends SymfonyExtension
 {