+++ tests/Resources/App/ZJKizaSqlTwigTestKernel.php
@@ -67,7 +67,6 @@ final class ZJKizaSqlTwigTestKernel extends Kernel
             'orm'  => [
                 'auto_generate_proxy_classes' => true,
                 'naming_strategy'             => 'doctrine.orm.naming_strategy.underscore_number_aware',
-                'auto_mapping'                => true,
                 'enable_lazy_ghost_objects'   => true,
                 'report_fields_where_declared' => true,
                 'mappings'                    => [